Add script to sniff Android device on the same network
Add a Python script that uses Scapy to sniff network traffic from an Android device on the same network. The script sets up a filter for the Android device's IP address and prints statistics about the packets received. 

Note: Make sure to update the <Android_Device_IP> in the code to the IP address of the Android machine you want to sniff.